+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ CHANGELOG
 * bug-fix: Append retry id to default Airflow job name to avoid name collisions in retry
 * bug-fix: Local Mode: No longer requires s3 permissions to run local entry point file
 * bug-fix: Local Mode: Move dependency on sagemaker_s3_output from rl.estimator to model
+* doc-fix: Fix quotes in estimator.py and model.py
 
 1.16.2
 ======